Move validators to own validator package (not inside pages)
