--- src/pages/validators/double-validator.ts
@@ -1,13 +1,0 @@
-import { FormControl, ValidationErrors } from '@angular/forms';
-
-export class DoubleValidator {
-
-  static isValid(control: FormControl): ValidationErrors | null {
-    const name = control.value;
-    if (isNaN(control.value)) {
-      return { 'not a number': { name } };
-    } else {
-      return null;
-    }
-  }
-}
